requirements.txt
absl-py==0.11.0
activation==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==1.5.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.10
h5py==2.10.0
html5lib==0.9999999
imageio==2.9.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
k==0.0.1
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.16.0
opencv-python==4.4.0.46
pandas==0.19.1
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.51.0
traitlets==4.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
